% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_helpers.R
\name{build_mixture_probabilities}
\alias{build_mixture_probabilities}
\title{Build the set of mixture combination prior probabilities}
\usage{
build_mixture_probabilities(x, y)
}
\arguments{
\item{x}{A vector of probabilities}

\item{y}{A matrix with each row representing a combination of
draws from binary random variables}
}
\description{
Build the set of mixture combination prior probabilities
}
