% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_norm_factors.R
\name{calc_norm_factors}
\alias{calc_norm_factors}
\title{Calculate fixed sample normalization factors for RNA-Seq
or Ribo-Seq data}
\usage{
calc_norm_factors(counts, method = "TMM", ref_column = NULL, ...)
}
\arguments{
\item{counts}{Matrix of read counts with one row per gene and one column
per sample}

\item{method}{Method to use}

\item{ref_column}{Column id of counts to use as reference}

\item{...}{Additional arguments passed to `method` function}
}
\description{
This is a lightweight re-implementation of the normalization factors calculation from the `edgeR` package
}
\seealso{
[https://code.bioconductor.org/browse/edgeR/blob/RELEASE_3_19/R/calcNormFactors.R]
}
