% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dte}
\alias{dte}
\title{Translational Efficiency Ribo-Seq and RNA-Seq data}
\format{
## `dte`
A data frame with 27,628 rows and 37 columns:
\describe{
  \item{geneid}{Gene identifier}
  \item{ribo_..., rna_...}{Ribo-Seq and RNA-Seq counts for treatment
  and control samples at 5 and 8 days post infection}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/bioproject/PRJNA950066>
}
\usage{
dte
}
\description{
Ribo-Seq and RNA-Seq counts per gene from
Report ...
}
\keyword{datasets}
