% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dte_stan.R
\name{dte_stan}
\alias{dte_stan}
\title{Bayesian hierarchical Poisson GLMM with discrete mixture prior for
detection of differential translational efficiency in RiboSeq studies}
\usage{
dte_stan(
  G,
  X_g,
  Z_g,
  y,
  a_p = 1,
  b_p = 2,
  a_sig2 = NULL,
  b_sig2 = NULL,
  sig2_mu = NULL,
  sig2_u = NULL,
  ...
)
}
\arguments{
\item{G}{Number of groups}

\item{X_g}{Fixed effects design matrix for a single group}

\item{Z_g}{Random effects design matrix for a single group}

\item{y}{Numeric vector of output values, sorted by group}

\item{a_p}{scale parameter for Beta prior on `p`}

\item{b_p}{shape parameter for Beta prior on `p`}

\item{a_sig2}{vector of shape parameters for Inverse Gamma priors on `sig2`}

\item{b_sig2}{vector of scale parameters for Inverse Gamma priors on `sig2`}

\item{sig2_mu}{vector of variances for the Gaussian prior on `mu`}

\item{sig2_u}{vector of variances for the Gaussian prior on random effects}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bayesian hierarchical Poisson GLMM with discrete mixture prior for
detection of differential translational efficiency in RiboSeq studies
}
