% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrast_posterior.R
\name{get_contrast_posterior}
\alias{get_contrast_posterior}
\title{Get per-draw probability that a given contrast is 0}
\usage{
get_contrast_posterior(contrast, comps, d_pmf, beta)
}
\arguments{
\item{contrast}{numeric vector giving the contrast}

\item{comps}{taken from `standata` list}

\item{d_pmf}{posterior draws for `d_mpf`}

\item{beta}{posterior draws for `beta`}
}
\description{
Get per-draw probability that a given contrast is 0
}
