% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{get_sim}
\alias{get_sim}
\title{Extract a single simulated data set from a simulation
run of the sampler}
\usage{
get_sim(sim_draws, draw = 1, chain = 1)
}
\arguments{
\item{sim_draws}{draws object from a simulation run}

\item{draw}{which draw to extract}

\item{chain}{which chain to extract from}
}
\description{
Extract a single simulated data set from a simulation
run of the sampler
}
