% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mov10}
\alias{mov10}
\title{Gene Expression for MOV10 data}
\format{
## `mov10`
A data frame with 14,675 rows and 9 columns:
\describe{
 \item{geneid}{Gene identifier}
 \item{low_..., high_..., control_...}{Gene expression counts
 for each sample}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE50499>
}
\usage{
mov10
}
\description{
Counts per gene
}
\keyword{datasets}
