% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dte_data.R
\name{prep_dte_data}
\alias{prep_dte_data}
\title{Construct required arguments to [dte_stan()] from
a data frame of Ribo-Seq and RNA-Seq read counts}
\usage{
prep_dte_data(counts, sample_names, X_g, Z_g)
}
\arguments{
\item{counts}{data frame of read counts, with one row per gene,
column geneid, and one column per sample
@param sample_names vector of sample column names in `counts`}

\item{X_g}{matrix representing the fixed effects design for
counts from a single gene}

\item{Z_g}{matrix representing the random effects design,
i.e. which Ribo-Seq and mRNA samples are paired,
for counts from a single gene}
}
\description{
Construct required arguments to [dte_stan()] from
a data frame of Ribo-Seq and RNA-Seq read counts
}
