% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dte_model.R
\name{run_dte_model}
\alias{run_dte_model}
\title{Fit the DTE model}
\usage{
run_dte_model(
  G,
  X_g,
  Z_g,
  y,
  norm_factors,
  a_p = 1,
  b_p = 2,
  a_sig2 = NULL,
  b_sig2 = NULL,
  sig2_mu = NULL,
  sig2_u = NULL,
  method = c("sampling", "vb", "pathfinder"),
  ...
)
}
\arguments{
\item{G}{Number of groups}

\item{X_g}{Fixed effects design matrix for a single group}

\item{Z_g}{Random effects design matrix for a single group}

\item{y}{Numeric vector of output values, sorted by group}

\item{norm_factors}{Numeric vector of normalization factors
for each sample}

\item{a_p}{scale parameter for Beta prior on `p`}

\item{b_p}{shape parameter for Beta prior on `p`}

\item{a_sig2}{vector of shape parameters for Inverse Gamma priors on `sig2`}

\item{b_sig2}{vector of scale parameters for Inverse Gamma priors on `sig2`}

\item{sig2_mu}{vector of variances for the Gaussian prior on `mu`}

\item{sig2_u}{vector of variances for the Gaussian prior on random effects}

\item{method}{One of `c("sampling", "vb", "pathfinder")`}

\item{...}{Named arguments to the `sample()` method of CmdStan model
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
An object of class `CmdStanMCMC`
}
\description{
Fit the DTE Stan model and return fit object
}
\concept{models}
