% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpl_glmm_mix_model.R
\name{run_hpl_glmm_mix_model}
\alias{run_hpl_glmm_mix_model}
\title{Fit the hpl_glmm_mix model}
\usage{
run_hpl_glmm_mix_model(
  standata,
  method = c("sample", "vb", "pathfinder"),
  run_estimation = FALSE,
  use_multithread = FALSE,
  grainsize = NULL,
  ...
)
}
\arguments{
\item{standata}{A list of data to be passed to the Stan model,
matching the format of the `standata` element of a `seqlist` object}

\item{method}{One of `c("sample", "vb", "pathfinder")`}

\item{run_estimation}{one of `c(0, 1)`; if 0, samples from the prior only
and ignores the data}

\item{use_multithread}{Use the multithread-enabled version of the Stan model?}

\item{grainsize}{Grainsize for multithread}

\item{...}{Named arguments to the `sample()` method of CmdStan model
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
An object of class `CmdStanMCMC`
}
\description{
Fit the hpl_glmm_mix Stan model and return fit object
}
\concept{models}
