% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpl_glmm_mix_multithread_model.R
\name{run_hpl_glmm_mix_mt_model}
\alias{run_hpl_glmm_mix_mt_model}
\title{Fit the hpl_glmm_mix_multithread model}
\usage{
run_hpl_glmm_mix_mt_model(
  method = c("sample", "vb", "pathfinder"),
  run_estimation = 0,
  grainsize = 1,
  G,
  X_g,
  Z_g,
  comps,
  prob,
  y = NULL,
  a_sig2 = NULL,
  b_sig2 = NULL,
  a_sig2_mu = NULL,
  b_sig2_mu = NULL,
  a_mu_offset = NULL,
  b_mu_offset = NULL,
  a_sig2_offset = NULL,
  b_sig2_offset = NULL,
  a_sig2_u = NULL,
  b_sig2_u = NULL,
  normfactors_known = FALSE,
  A_S = NULL,
  B_S = NULL,
  S_DATA = NULL,
  ...
)
}
\arguments{
\item{method}{One of `c("sample", "vb", "pathfinder")`}

\item{run_estimation}{one of `c(0, 1)`; if 0, samples from the prior only
and ignores the data}

\item{grainsize}{Grainsize for multithread}

\item{G}{Number of groups}

\item{X_g}{Fixed effects design matrix for a single group}

\item{Z_g}{Random effects design matrix for a single group}

\item{comps}{Matrix encoding bernoulli mixture priors}

\item{prob}{Vector giving the prior probability of each combination of
bernoulli components, i.e. of each row in `comps`}

\item{y}{Numeric vector of output values, sorted by group}

\item{a_sig2}{vector of shape parameters
for Inverse Gamma priors on `sig2`}

\item{b_sig2}{vector of scale parameters
for Inverse Gamma priors on `sig2`}

\item{a_sig2_mu}{vector of shape parameters
for Inverse Gamma priors on `sig2_mu`}

\item{b_sig2_mu}{vector of scale parameters
for Inverse Gamma priors on `sig2_mu`}

\item{a_mu_offset}{vector of shape parameters
for Inverse Gamma priors on `mu_offset`}

\item{b_mu_offset}{vector of scale parameters
for Inverse Gamma priors on `mu_offset`}

\item{a_sig2_offset}{vector of shape parameters
for Inverse Gamma priors on `sig2_offset`}

\item{b_sig2_offset}{vector of scale parameters
for Inverse Gamma priors on `sig2_offset`}

\item{a_sig2_u}{vector of shape parameters
for Inverse Gamma priors on `sig2_u`}

\item{b_sig2_u}{vector of scale parameters
for Inverse Gamma priors on `sig2_u`}

\item{normfactors_known}{Use fixed normalization factors extrinsic to
the model?}

\item{A_S}{location parameter for `S_PARAM` if `!normfactors_known`}

\item{B_S}{scale parameter for `S_PARAM` if `!normfactors_known`}

\item{S_DATA}{Numeric vector of normalization factors
for each sample; optional, if `NULL` and `normfactors_known == TRUE`,
normalization factors will be estimated by the `TMM` method as
described in the `{edgeR}` package}

\item{...}{Named arguments to the `sample()` method of CmdStan model
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
An object of class `CmdStanMCMC`
}
\description{
Fit the hpl_glmm_mix Stan model and return fit object
}
\seealso{
Other models: 
\code{\link{run_hpl_glmm_mix_model}()}
}
\concept{models}
