% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_standata.R
\name{set_standata}
\alias{set_standata}
\title{Set the value of one of the modifiable elements of `standata`}
\usage{
set_standata(
  standata,
  a_sig2 = NULL,
  b_sig2 = NULL,
  a_sig2_mu = NULL,
  b_sig2_mu = NULL,
  a_mu_offset = NULL,
  b_mu_offset = NULL,
  a_sig2_offset = NULL,
  b_sig2_offset = NULL,
  a_sig2_u = NULL,
  b_sig2_u = NULL,
  A_S = NULL,
  B_S = NULL,
  S_DATA = NULL,
  normfactors_known = NULL
)
}
\arguments{
\item{standata}{The data list to be updated}

\item{a_sig2}{vector of shape parameters
for Inverse Gamma priors on `sig2`}

\item{b_sig2}{vector of scale parameters
for Inverse Gamma priors on `sig2`}

\item{a_sig2_mu}{vector of shape parameters
for Inverse Gamma priors on `sig2_mu`}

\item{b_sig2_mu}{vector of scale parameters
for Inverse Gamma priors on `sig2_mu`}

\item{a_mu_offset}{vector of shape parameters
for Inverse Gamma priors on `mu_offset`}

\item{b_mu_offset}{vector of scale parameters
for Inverse Gamma priors on `mu_offset`}

\item{a_sig2_offset}{vector of shape parameters
for Inverse Gamma priors on `sig2_offset`}

\item{b_sig2_offset}{vector of scale parameters
for Inverse Gamma priors on `sig2_offset`}

\item{a_sig2_u}{vector of shape parameters
for Inverse Gamma priors on `sig2_u`}

\item{b_sig2_u}{vector of scale parameters
for Inverse Gamma priors on `sig2_u`}

\item{A_S}{location parameter for `S_PARAM` if `!normfactors_known`}

\item{B_S}{scale parameter for `S_PARAM` if `!normfactors_known`}

\item{S_DATA}{Numeric vector of normalization factors
for each sample; optional, if `NULL` and `normfactors_known == TRUE`,
normalization factors will be estimated by the `TMM` method as
described in the `{edgeR}` package}

\item{normfactors_known}{Use fixed normalization factors extrinsic to
the model?}

\item{beta_phi_prior}{2-vector giving location and scale parameters for
the prior distribution of `beta_phi`}
}
\description{
Set the value of one of the modifiable elements of `standata`
}
