% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuch_tumor}
\alias{tuch_tumor}
\title{Tuch tumor data}
\format{
## `tuch_tumor`
A data frame with 15,668 rows and 9 columns:
\describe{
 \item{id_ref_seq}{Gene identifier}
 \item{name_of_gene}{Gene name}
 \item{number_of_exons}{Exons per gene}
 \item{8n, 8t, 33n, 33t, 51n, 51t}{Gene expression counts
 for each sample}
}
}
\source{
<https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0009317#s5> # nolint
}
\usage{
tuch_tumor
}
\description{
Counts per gene
}
\keyword{datasets}
