% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_helpers.R
\name{build_mixture_indicator}
\alias{build_mixture_indicator}
\title{Build the set of mixture prior indicators}
\usage{
build_mixture_indicator(x)
}
\arguments{
\item{x}{An indicator vector identifying which
regression parameters should use the mixture prior}
}
\description{
Build the set of mixture prior indicators
}
