% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpl_glmm_mix_model.R
\name{run_hpl_glmm_mix_model}
\alias{run_hpl_glmm_mix_model}
\title{Fit the hpl_glmm_mix model}
\usage{
run_hpl_glmm_mix_model(
  method = c("sampling", "vb", "pathfinder"),
  run_estimation = 0,
  G,
  X_g,
  Z_g,
  comps,
  prob,
  y = NULL,
  a_sig2 = NULL,
  b_sig2 = NULL,
  a_sig2_mu = NULL,
  b_sig2_mu = NULL,
  a_mu_offset = NULL,
  b_mu_offset = NULL,
  a_sig2_offset = NULL,
  b_sig2_offset = NULL,
  a_sig2_u = NULL,
  b_sig2_u = NULL,
  S = NULL,
  ...
)
}
\arguments{
\item{method}{One of `c("sampling", "vb", "pathfinder")`}

\item{G}{Number of groups}

\item{X_g}{Fixed effects design matrix for a single group}

\item{Z_g}{Random effects design matrix for a single group}

\item{y}{Numeric vector of output values, sorted by group}

\item{a_sig2}{vector of shape parameters for Inverse Gamma priors on `sig2`}

\item{b_sig2}{vector of scale parameters for Inverse Gamma priors on `sig2`}

\item{S}{Numeric vector of normalization factors
for each sample}

\item{...}{Named arguments to the `sample()` method of CmdStan model
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
An object of class `CmdStanMCMC`
}
\description{
Fit the hpl_glmm_mix Stan model and return fit object
}
\seealso{
Other models: 
\code{\link{run_dte_model}()}
}
\concept{models}
