% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqlist.R
\name{seqlist}
\alias{seqlist}
\title{R6 class representing data from an RNA-Seq or Ribo-Seq experiment}
\description{
A `seqlist` has data from the experiment and information about
the experimental design and model specification.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{counts}}{Numeric matrix of output value, where rows represent genes
or tags and columns represent samples}

\item{\code{tags}}{A dataframe of tag metadata, matching the row-order of `counts`}

\item{\code{fixed_design}}{Fixed effects design matrix for a single group}

\item{\code{random_design}}{Random effects design matrix for a single group}

\item{\code{mixture_probabilities}}{Vector giving the prior probabilities that
each regression parameter is drawn from the 0-component of the mixture
prior}

\item{\code{standata}}{List of data to be passed to Stan}

\item{\code{fit}}{Fit object returned by `run_model()`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-seqlist-new}{\code{seqlist$new()}}
\item \href{#method-seqlist-set_counts}{\code{seqlist$set_counts()}}
\item \href{#method-seqlist-set_fixed_design}{\code{seqlist$set_fixed_design()}}
\item \href{#method-seqlist-set_random_design}{\code{seqlist$set_random_design()}}
\item \href{#method-seqlist-set_mixture_probabilities}{\code{seqlist$set_mixture_probabilities()}}
\item \href{#method-seqlist-initialize_standata}{\code{seqlist$initialize_standata()}}
\item \href{#method-seqlist-set_standata}{\code{seqlist$set_standata()}}
\item \href{#method-seqlist-run_model}{\code{seqlist$run_model()}}
\item \href{#method-seqlist-clone}{\code{seqlist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-new"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-new}{}}}
\subsection{Method \code{new()}}{
Create a new `seqlist`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$new(
  counts = NULL,
  tags = NULL,
  fixed_design = NULL,
  random_design = NULL,
  mixture_probabilities = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{counts}}{Value for `counts` field}

\item{\code{tags}}{Value for `tags` field}

\item{\code{fixed_design}}{Value for `fixed_design` field}

\item{\code{random_design}}{Value for `random_design` field}

\item{\code{mixture_probabilities}}{Value for `mixture_probabilities` field}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `seqlist` object
set the `counts` field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-set_counts"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-set_counts}{}}}
\subsection{Method \code{set_counts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$set_counts(counts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{counts}}{the counts to use
set the `fixed_design` field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-set_fixed_design"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-set_fixed_design}{}}}
\subsection{Method \code{set_fixed_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$set_fixed_design(fixed_design)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fixed_design}}{the fixed design matrix to use
set the `random_design` matrix field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-set_random_design"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-set_random_design}{}}}
\subsection{Method \code{set_random_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$set_random_design(random_design)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{random_design}}{the random design matrix to use
set the `mixture_probabilities` field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-set_mixture_probabilities"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-set_mixture_probabilities}{}}}
\subsection{Method \code{set_mixture_probabilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$set_mixture_probabilities(mixture_probabilities)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mixture_probabilities}}{the vector of mixture probabilities
set the `standata` field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-initialize_standata"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-initialize_standata}{}}}
\subsection{Method \code{initialize_standata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$initialize_standata(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments passed to private methods
Set the value of one of the modifiable elements of `standata`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-set_standata"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-set_standata}{}}}
\subsection{Method \code{set_standata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$set_standata(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments passed to `ngstan::set_standata()`
run the model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-run_model"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-run_model}{}}}
\subsection{Method \code{run_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$run_model(
  method = c("sample", "vb", "pathfinder"),
  run_estimation = FALSE,
  use_multithread = FALSE,
  grainsize = NULL,
  modify_in_place = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{the sampling algorithm to use}

\item{\code{run_estimation}}{include the likelihood in the objective?}

\item{\code{use_multithread}}{use the multithread-enabled model?}

\item{\code{grainsize}}{grainsize for multithread}

\item{\code{modify_in_place}}{add the fit to the `seqlist` object? otherwise
return it}

\item{\code{...}}{arguments passed to the `{cmdstanr}` function
identified by `method`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-seqlist-clone"></a>}}
\if{latex}{\out{\hypertarget{method-seqlist-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{seqlist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
